Introduced RangeConfigurator and removed configure hook for Arbitrary providers
