+++ src/main/java/net/jqwik/providers/BigDecimalArbitraryProvider.java
@@ -1,13 +1,11 @@
 package net.jqwik.providers;
 
-import net.jqwik.api.*;
-import net.jqwik.api.arbitraries.*;
-import net.jqwik.api.constraints.*;
-import net.jqwik.api.providers.*;
-
 import java.math.*;
 import java.util.*;
 import java.util.function.*;
+
+import net.jqwik.api.*;
+import net.jqwik.api.providers.*;
 
 public class BigDecimalArbitraryProvider implements ArbitraryProvider {
 	@Override
@@ -19,9 +17,4 @@ public class BigDecimalArbitraryProvider implements ArbitraryProvider {
 	public Arbitrary<?> provideFor(GenericType targetType, Function<GenericType, Optional<Arbitrary<?>>> subtypeProvider) {
 		return Arbitraries.bigDecimals();
 	}
-
-	public BigDecimalArbitrary configure(BigDecimalArbitrary arbitrary, DoubleRange range) {
-		return arbitrary.withMin(new BigDecimal(range.min())).withMax(new BigDecimal(range.max()));
-	}
-
 }
