+++ src/main/java/net/jqwik/providers/BigIntegerArbitraryProvider.java
@@ -5,8 +5,6 @@ import java.util.*;
 import java.util.function.*;
 
 import net.jqwik.api.*;
-import net.jqwik.api.arbitraries.*;
-import net.jqwik.api.constraints.*;
 import net.jqwik.api.providers.*;
 
 public class BigIntegerArbitraryProvider implements ArbitraryProvider {
@@ -19,9 +17,4 @@ public class BigIntegerArbitraryProvider implements ArbitraryProvider {
 	public Arbitrary<?> provideFor(GenericType targetType, Function<GenericType, Optional<Arbitrary<?>>> subtypeProvider) {
 		return Arbitraries.bigIntegers();
 	}
-
-	public BigIntegerArbitrary configure(BigIntegerArbitrary arbitrary, LongRange range) {
-		return arbitrary.withMin(BigInteger.valueOf(range.min())).withMax(BigInteger.valueOf(range.max()));
-	}
-
 }
