+++ src/main/java/net/jqwik/providers/IntegerArbitraryProvider.java
@@ -1,12 +1,10 @@
 package net.jqwik.providers;
 
-import net.jqwik.api.*;
-import net.jqwik.api.arbitraries.*;
-import net.jqwik.api.constraints.*;
-import net.jqwik.api.providers.*;
-
 import java.util.*;
 import java.util.function.*;
+
+import net.jqwik.api.*;
+import net.jqwik.api.providers.*;
 
 public class IntegerArbitraryProvider implements ArbitraryProvider {
 	@Override
@@ -18,9 +16,4 @@ public class IntegerArbitraryProvider implements ArbitraryProvider {
 	public Arbitrary<?> provideFor(GenericType targetType, Function<GenericType, Optional<Arbitrary<?>>> subtypeProvider) {
 		return Arbitraries.integers();
 	}
-
-	public IntegerArbitrary configure(IntegerArbitrary arbitrary, IntRange range) {
-		return arbitrary.withMin(range.min()).withMax(range.max());
-	}
-
 }
