+++ src/main/java/net/jqwik/providers/ShortArbitraryProvider.java
@@ -4,8 +4,6 @@ import java.util.*;
 import java.util.function.*;
 
 import net.jqwik.api.*;
-import net.jqwik.api.arbitraries.*;
-import net.jqwik.api.constraints.*;
 import net.jqwik.api.providers.*;
 
 public class ShortArbitraryProvider implements ArbitraryProvider {
@@ -18,9 +16,4 @@ public class ShortArbitraryProvider implements ArbitraryProvider {
 	public Arbitrary<?> provideFor(GenericType targetType, Function<GenericType, Optional<Arbitrary<?>>> subtypeProvider) {
 		return Arbitraries.shorts();
 	}
-
-	public ShortArbitrary configure(ShortArbitrary arbitrary, ShortRange range) {
-		return arbitrary.withMin(range.min()).withMax(range.max());
-	}
-
 }
