+++ src/main/resources/META-INF/services/net.jqwik.api.providers.ArbitraryConfigurator
@@ -3,3 +3,4 @@ net.jqwik.configurators.ScaleConfigurator
 net.jqwik.configurators.SizeConfigurator
 net.jqwik.configurators.CharsConfigurator
 net.jqwik.configurators.StringLengthConfigurator
+net.jqwik.configurators.RangeConfigurator
